Enable existing optimizations in the kernel source
This commit enables features that have been integrated into the kernel source but were not enabled, thereby enhancing system performance and user experience. These configurations are not made casually; they were integrated into the kernel before the initial upload of the kernel source, and this commit only serves to turn on the switches!

1. Provide necessary filesystem support for the Mountify module to enhance system security and compatibility.
2. Enable compiler optimizations to improve kernel runtime efficiency and system response speed.
3. Add high-performance memory compression and data compression algorithms to enhance memory utilization and data processing speed.
4. Integrate various advanced TCP congestion control algorithms and set BBR as the default option to optimize network transmission performance.
5. Port and enable the IO scheduler optimized for Samsung devices to improve storage I/O performance and read/write speeds.
6. Add kernel-level dynamic recompilation support to provide underlying support for system functions such as application freezing/tombstone.

Again, these configuration options are not enabled casually; they were integrated into the kernel before the initial upload of the kernel source, and this commit only serves to turn on the switches for these functions!

Signed-off-by: Winkmoon <3432632982@qq.com>